Add Vuex modules for events and user information
