+++ src/store/modules/user.js
@@ -1,0 +1,6 @@
+export const state = {
+    user: {
+        id: 'abc123',
+        name: 'Adam'
+    }
+}
