+++ src/views/EventCreate.vue
@@ -71,7 +71,7 @@ export default {
     methods: {
         async createEvent() {
             try {
-                await this.$store.dispatch('createEvent', this.event)
+                await this.$store.dispatch('event/createEvent', this.event)
                 this.$router.push({
                     name: 'event-show',
                     params: {
@@ -84,7 +84,7 @@ export default {
             }
         },
         createFreshEventObject() {
-            const user = this.$store.state.user
+            const user = this.$store.state.user.user
             const id = Math.floor(Math.random() * 10000000)
 
             return {
