+++ src/views/EventList.vue
@@ -1,7 +1,11 @@
 <template>
     <div>
-        <h1>Events Listing</h1>
-        <EventCard v-for="event in events" :key="event.id" :event="event" />
+        <h1>Events for {{ user.user.name }}</h1>
+        <EventCard
+            v-for="event in event.events"
+            :key="event.id"
+            :event="event"
+        />
         <template v-if="page != 1">
             <router-link
                 :to="{ name: 'event-list', query: { page: page - 1 } }"
@@ -10,7 +14,7 @@
                 Prev Page
             </router-link>
             <router-link
-                v-if="eventsTotal > page * 2"
+                v-if="hasNextPage"
                 :to="{ name: 'event-list', query: { page: page + 1 } }"
                 rel="next"
             >
@@ -30,18 +34,22 @@ export default {
         EventCard
     },
     computed: {
-        ...mapState(['events', 'eventsTotal']),
         page() {
             return parseInt(this.$route.query.page) || 1
-        }
+        },
+        hasNextPage() {
+            return this.eventsTotal > this.page * this.perPage
+        },
+        ...mapState(['event', 'user'])
     },
     created() {
+        this.perPage = 2
         this.fetchEvents()
     },
     methods: {
         fetchEvents() {
-            this.$store.dispatch('fetchEvents', {
-                perPage: 2,
+            this.$store.dispatch('event/fetchEvents', {
+                perPage: this.perPage,
                 page: this.page
             })
         }
