+++ src/views/EventShow.vue
@@ -33,7 +33,7 @@
 </template>
 
 <script>
-import { mapState } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 
 export default {
     name: 'EventShow',
@@ -44,10 +44,15 @@ export default {
         }
     },
     computed: {
-        ...mapState(['event'])
+        ...mapState({
+            event: state => state.event.event
+        })
     },
     created() {
-        this.$store.dispatch('fetchEvent', this.id)
+        this.fetchEvent(this.id)
+    },
+    methods: {
+        ...mapActions('event', ['fetchEvent'])
     }
 }
 </script>
